Rename throwIfUnsuccessful
