+++ libs/repositories/src/UsdRepository/UsdRepositoryCoingecko.ts
@@ -2,7 +2,7 @@ import { injectable } from 'inversify';
 import { PricePoint, PriceStrategy, UsdRepository } from './UsdRepository';
 import { COINGECKO_PLATFORMS, coingeckoProClient } from '../coingecko';
 import { SupportedChainId } from '../types';
-import { throwIfUnsuccessful } from '../utils/fetch';
+import { throwIfUnsuccessful } from '../utils/throwIfUnsuccessful';
 
 /**
  * Number of days of data to fetch for each price strategy
